hooks: split player orchestration modules
