+++ crates/hooks/src/use_player_task.rs
@@ -10,6 +10,9 @@ use std::sync::Arc;
 use discord_presence::Presence;
 use discord_presence::cover_art;
 
+#[path = "player_task_event_pump.rs"]
+mod event_pump;
+
 #[cfg(target_os = "macos")]
 use player::systemint::set_background_handler;
 
@@ -176,16 +179,7 @@ fn nudge_event_loop() {
 }
 
 pub fn use_player_task(ctrl: PlayerController) {
-    let presence: Option<Arc<Presence>> = use_context();
-    let mut config: Signal<AppConfig> = use_context();
-
-    let mut last_title = use_signal(String::new);
-    let mut last_source: Signal<Option<String>> = use_signal(|| None);
-    let mut was_playing = use_signal(|| false);
-    let mut discord_cover_url: Signal<Option<String>> = use_signal(|| None);
-    let mut discord_cover_resolving_for = use_signal(String::new);
-    let mut discord_cover_sent = use_signal(|| false);
-
+    let config: Signal<AppConfig> = use_context();
     #[cfg(target_os = "macos")]
     use_hook(move || {
         init_bg_channel();
@@ -723,597 +717,5 @@ pub fn use_player_task(ctrl: PlayerController) {
         spotify_pump.set(Some(task));
     });
 
-    let gens = crate::db_reactivity::use_generations();
-    use_future(move || {
-        let mut ctrl = ctrl;
-        let presence = presence.clone();
-        let mut last_ping = std::time::Instant::now();
-        let mut last_progress_report = std::time::Instant::now();
-        let mut last_discord_enabled = false;
-        let mut last_jellyfin_id: Option<String> = None;
-        #[cfg(target_os = "macos")]
-        let mut last_now_playing_refresh = std::time::Instant::now();
-        let mut last_lyrics_prefetch_track: Option<String> = None;
-
-        async move {
-            let mut last_progress_secs: u64 = u64::MAX;
-            let mut prev_playing = false;
-            let mut last_recent_path: Option<String> = None;
-            let bg_notify = BG_NOTIFY.get_or_init(tokio::sync::Notify::new);
-            // Engine events (Position once per second while playing, Ended,
-            // phase changes) wake this loop; the coarse timer only covers the
-            // genuinely periodic work (Jellyfin reports, Discord, refreshes).
-            let mut engine_events = ctrl.player.peek().subscribe();
-            loop {
-                let mut woke_event = None;
-                tokio::select! {
-                    _ = bg_notify.notified() => {},
-                    _ = tokio::time::sleep(std::time::Duration::from_millis(1000)) => {},
-                    event = engine_events.recv() => { woke_event = event; },
-                }
-                // Drain whatever queued while we were asleep. Progress and
-                // completion are read from the live status snapshot below;
-                // only phase transitions need the events themselves.
-                while let Some(event) = woke_event.take().or_else(|| engine_events.try_recv().ok())
-                {
-                    use player::engine::{Event, Phase};
-                    match event {
-                        Event::PhaseChanged {
-                            token,
-                            phase: phase @ (Phase::Playing | Phase::Paused),
-                        } => {
-                            if token == ctrl.intent.peek().token() {
-                                ctrl.is_playing.set(phase == Phase::Playing);
-                            } else if phase == Phase::Playing
-                                && ctrl.player.peek().session_token() == token
-                            {
-                                // A session we no longer intend is audibly live —
-                                // stop it. (Guarded on session_token so an event
-                                // outrun by our own revert seek is ignored.)
-                                ctrl.player.peek().stop_for_transition();
-                            }
-                        }
-                        // No session — nothing audible. Ended is left to the
-                        // auto-advance check below, which needs is_playing.
-                        Event::PhaseChanged {
-                            phase: Phase::Idle, ..
-                        } if !*ctrl.external_active.peek() => ctrl.is_playing.set(false),
-                        // Commit the deferred crossfade UI exactly on fade end.
-                        Event::TrackSwitched { token, .. } => {
-                            ctrl.commit_transition(token);
-                            last_progress_secs = u64::MAX;
-                        }
-                        // A promoted load we superseded/cancelled started (the
-                        // end-of-queue race): stop it. Same session_token guard.
-                        Event::Loaded { token }
-                            if token != ctrl.intent.peek().token()
-                                && ctrl.player.peek().session_token() == token =>
-                        {
-                            ctrl.player.peek().stop_for_transition();
-                        }
-                        // Device lost on a live session (radio unplug): the load
-                        // reply can't report it. Banner + stop, no auto-reconnect.
-                        Event::Error { token, message } if token == ctrl.intent.peek().token() => {
-                            tracing::warn!(%message, "engine reported a playback error");
-                            ctrl.fail_load(token, &message);
-                            ctrl.is_playing.set(false);
-                        }
-                        _ => {}
-                    }
-                }
-
-                nudge_event_loop();
-
-                let cmds = drain_bg_cmds();
-                let last_seek = cmds.iter().rposition(|c| matches!(c, BgCmd::Seek(_)));
-                for (i, cmd) in cmds.into_iter().enumerate() {
-                    match cmd {
-                        BgCmd::Play => ctrl.resume(),
-                        BgCmd::Pause => ctrl.pause(),
-                        BgCmd::Toggle => ctrl.toggle(),
-                        BgCmd::Next => ctrl.play_next(),
-                        BgCmd::Prev => ctrl.play_prev(),
-                        BgCmd::Seek(secs) if Some(i) == last_seek => {
-                            let pos = std::time::Duration::try_from_secs_f64(secs)
-                                .unwrap_or(std::time::Duration::ZERO);
-                            ctrl.seek(pos);
-                        }
-                        BgCmd::Seek(_) => {}
-                        #[cfg(target_os = "android")]
-                        BgCmd::ToggleShuffle => {
-                            let on = *ctrl.shuffle.peek();
-                            ctrl.set_shuffle(!on);
-                        }
-                        #[cfg(target_os = "android")]
-                        BgCmd::CycleRepeat => {
-                            let next = ctrl.loop_mode.peek().next();
-                            ctrl.set_loop_mode(next);
-                        }
-                    }
-                }
-
-                let is_playing = *ctrl.is_playing.read();
-                let external = *ctrl.external_active.peek();
-
-                {
-                    let current_track = {
-                        let idx = *ctrl.current_queue_index.read();
-                        ctrl.get_track_at(idx)
-                    };
-                    if let Some(track) = current_track
-                        && is_playing
-                    {
-                        let uid = track.id.uid().to_string();
-                        if last_recent_path.as_ref() != Some(&uid) {
-                            last_recent_path = Some(uid);
-                            // Records under the active source's partition — same
-                            // source the rest of now-playing resolves through.
-                            let key = track.id.key().into_owned();
-                            let source = ctrl.active_source.peek().clone();
-                            spawn(async move {
-                                if source.record_recent(&key).await.is_ok() {
-                                    gens.bump(crate::db_reactivity::Table::Recents);
-                                }
-                            });
-                        }
-                    }
-                }
-
-                let lyrics_prefetch = {
-                    let current_idx = *ctrl.current_queue_index.read();
-                    ctrl.get_track_at(current_idx + 1)
-                };
-
-                if let Some(next_track) = lyrics_prefetch {
-                    let next_track_key = next_track.id.uid().to_string();
-                    // Radio has no lyrics to prefetch.
-                    if !crate::playback_ref::PlaybackItemRef::parse(&next_track_key).is_radio()
-                        && last_lyrics_prefetch_track.as_ref() != Some(&next_track_key)
-                    {
-                        last_lyrics_prefetch_track = Some(next_track_key);
-                        let (
-                            server_url,
-                            server_token,
-                            server_user_id,
-                            prefer_local,
-                            enable_musixmatch,
-                        ) = {
-                            let conf = config.read();
-                            let prefer_local = conf.prefer_local_lyrics;
-                            let enable_musixmatch = conf.enable_musixmatch_lyrics;
-                            if let Some(server) = &conf.server {
-                                (
-                                    Some(server.url.clone()),
-                                    server.access_token.clone(),
-                                    server.user_id.clone(),
-                                    prefer_local,
-                                    enable_musixmatch,
-                                )
-                            } else {
-                                (None, None, None, prefer_local, enable_musixmatch)
-                            }
-                        };
-
-                        spawn(async move {
-                            let next_track_path = next_track.id.uid();
-                            let mut lyrics_request = utils::lyrics::LyricsRequest::new(
-                                next_track.artist,
-                                next_track.title,
-                                next_track.album,
-                                next_track.duration,
-                                &next_track_path,
-                            )
-                            .with_server(
-                                server_url.as_deref(),
-                                server_token.as_deref(),
-                                server_user_id.as_deref(),
-                            )
-                            .prefer_local(prefer_local)
-                            .enable_musixmatch(enable_musixmatch);
-                            // Lazily attach Apple Music auth for the lyrics provider.
-                            if next_track_path.starts_with("applemusic:") {
-                                let am_auth = config.read().server.as_ref().and_then(|server| {
-                                    if server.service != MusicService::AppleMusic {
-                                        return None;
-                                    }
-                                    let token = server.access_token.clone()?;
-                                    let catalog_id = next_track_path
-                                        .strip_prefix("applemusic:")
-                                        .unwrap_or(&next_track_path)
-                                        .to_string();
-                                    Some(utils::lyrics::AppleMusicLyricsAuth {
-                                        token,
-                                        bearer_token: String::new(),
-                                        storefront: server.apple_music_storefront.clone(),
-                                        language: server.apple_music_language.clone(),
-                                        catalog_id,
-                                    })
-                                });
-                                if let Some(mut auth) = am_auth {
-                                    if let Ok(bt) =
-                                        ::server::applemusic::auth::get_bearer_token().await
-                                    {
-                                        auth.bearer_token = bt;
-                                    }
-                                    lyrics_request = lyrics_request.apple_music_auth(auth);
-                                }
-                            }
-                            let _ = utils::lyrics::fetch_lyrics_for_request(&lyrics_request).await;
-                        });
-                    }
-                }
-
-                // Android has no Discord; force-disable so the cover-art resolution and
-                // presence updates below are all skipped (the Presence context is None too).
-                #[cfg(target_os = "android")]
-                let discord_enabled = false;
-                #[cfg(not(target_os = "android"))]
-                let discord_enabled = config.read().discord_presence.unwrap_or(true);
-                let discord_paused_enabled = config.read().discord_presence_paused.unwrap_or(true);
-                let discord_source_name = config
-                    .read()
-                    .discord_presence_source
-                    .unwrap_or(true)
-                    .then(|| {
-                        config
-                            .read()
-                            .active_service()
-                            .map_or("Local", |s| s.display_name())
-                    });
-                let source_changed = last_source.peek().as_deref() != discord_source_name;
-
-                // Discord may start after Kopuz or restart mid-session: retry a
-                // dropped IPC connection and restore the last activity once it's up.
-                if discord_enabled && let Some(ref p) = presence {
-                    p.tick();
-                }
-
-                let pos = ctrl.player.read().get_position();
-                let mut defer_player_progress = false;
-
-                // Deferred crossfade UI: show the outgoing track's live position
-                // and hold back the incoming one until TrackSwitched commits.
-                if ctrl.pending_crossfade_ui.peek().is_some() {
-                    if let Some(fading) = ctrl.player.read().fading_position() {
-                        ctrl.current_song_progress.set(fading.as_secs());
-                    }
-                    defer_player_progress = true;
-                }
-
-                let jellyfin_info = {
-                    let conf = config.read();
-                    conf.server.clone().map(|s| (s, conf.device_id.clone()))
-                };
-
-                if let Some((server, _device_id)) = jellyfin_info
-                    && server.service == MusicService::Jellyfin
-                {
-                    // Session reporting goes through the active source (Jellyfin
-                    // overrides keepalive/report_*; others no-op). Reports are
-                    // gated to ≥5s/track-change, so resolving a fresh source per
-                    // report is cheap — no client cache needed.
-                    if last_ping.elapsed().as_secs() >= 30 {
-                        let source = ctrl.active_source.peek().clone();
-                        spawn(
-                            async move {
-                                let _ = source.keepalive().await;
-                            }
-                            .instrument(tracing::info_span!("jellyfin.keepalive")),
-                        );
-                        last_ping = std::time::Instant::now();
-                    }
-
-                    let track = {
-                        let current_idx = *ctrl.current_queue_index.read();
-                        ctrl.get_track_at(current_idx)
-                    };
-
-                    if let Some(track) = track {
-                        if let Some(current_id) = track
-                            .id
-                            .service()
-                            .filter(|s| *s == MusicService::Jellyfin)
-                            .map(|_| track.id.key().into_owned())
-                        {
-                            if last_jellyfin_id.as_ref() != Some(&current_id) {
-                                if let Some(old_id) = last_jellyfin_id {
-                                    let source = ctrl.active_source.peek().clone();
-                                    let ticks = pos.as_micros() as u64 * 10;
-                                    spawn(
-                                        async move {
-                                            let _ = source
-                                                .report_playback_stopped(&old_id, ticks)
-                                                .await;
-                                        }
-                                        .instrument(tracing::info_span!("playback.report")),
-                                    );
-                                }
-                                let source = ctrl.active_source.peek().clone();
-                                let current_id_clone = current_id.clone();
-                                spawn(
-                                    async move {
-                                        let _ =
-                                            source.report_playback_start(&current_id_clone).await;
-                                    }
-                                    .instrument(tracing::info_span!("playback.report")),
-                                );
-                                last_jellyfin_id = Some(current_id.clone());
-                            }
-
-                            if last_progress_report.elapsed().as_secs() >= 5
-                                || is_playing != prev_playing
-                            {
-                                let ticks = pos.as_micros() as u64 * 10;
-                                let source = ctrl.active_source.peek().clone();
-                                let current_id_clone = current_id.clone();
-                                spawn(
-                                    async move {
-                                        let _ = source
-                                            .report_playback_progress(
-                                                &current_id_clone,
-                                                ticks,
-                                                !is_playing,
-                                            )
-                                            .await;
-                                    }
-                                    .instrument(tracing::info_span!("playback.report")),
-                                );
-                                last_progress_report = std::time::Instant::now();
-                            }
-                        } else if let Some(old_id) = last_jellyfin_id.take() {
-                            let source = ctrl.active_source.peek().clone();
-                            let ticks = pos.as_micros() as u64 * 10;
-                            spawn(
-                                async move {
-                                    let _ = source.report_playback_stopped(&old_id, ticks).await;
-                                }
-                                .instrument(tracing::info_span!("playback.report")),
-                            );
-                        }
-                    } else if let Some(old_id) = last_jellyfin_id.take() {
-                        let source = ctrl.active_source.peek().clone();
-                        let ticks = pos.as_micros() as u64 * 10;
-                        spawn(
-                            async move {
-                                let _ = source.report_playback_stopped(&old_id, ticks).await;
-                            }
-                            .instrument(tracing::info_span!("playback.report")),
-                        );
-                    }
-                }
-
-                #[cfg(target_os = "macos")]
-                if last_now_playing_refresh.elapsed().as_secs() >= 10 {
-                    player::systemint::refresh_now_playing();
-                    last_now_playing_refresh = std::time::Instant::now();
-                }
-
-                if is_playing {
-                    let duration = *ctrl.current_song_duration.read();
-                    let pos_secs = if external {
-                        ctrl.displayed_progress_secs_f64() as u64
-                    } else {
-                        pos.as_secs()
-                    }
-                    .min(duration);
-                    let current_token = *ctrl.current_token.read();
-                    if !defer_player_progress && pos_secs != last_progress_secs {
-                        last_progress_secs = pos_secs;
-                        ctrl.current_song_progress.set(pos_secs);
-                    }
-
-                    if let Some(ref p) = presence {
-                        let title = ctrl.current_song_title.read().clone();
-                        let artist = ctrl.current_song_artist.read().clone();
-                        let album = ctrl.current_song_album.read().clone();
-                        let duration = *ctrl.current_song_duration.read();
-                        let progress = if duration == u64::MAX {
-                            0
-                        } else {
-                            pos.as_secs()
-                        };
-
-                        let song_key = format!("{}|{}|{}", title, artist, album);
-
-                        if discord_enabled && song_key != *discord_cover_resolving_for.peek() {
-                            discord_cover_resolving_for.set(song_key.clone());
-                            discord_cover_url.set(None);
-                            discord_cover_sent.set(false);
-
-                            let current_track = {
-                                let idx = *ctrl.current_queue_index.read();
-                                ctrl.get_track_at(idx)
-                            };
-                            if let Some(url) = current_track
-                                .as_ref()
-                                .and_then(cover_art::youtube_cover_art_url)
-                            {
-                                discord_cover_url.set(Some(url));
-                            } else {
-                                let mbid = current_track
-                                    .and_then(|track| track.musicbrainz_release_id.clone());
-                                let artist_c = artist.clone();
-                                let album_c = album.clone();
-                                let song_key_for_spawn = song_key.clone();
-                                spawn(
-                                    async move {
-                                        let resolved = cover_art::resolve_cover_art_url_cached(
-                                            mbid.as_deref(),
-                                            &artist_c,
-                                            &album_c,
-                                        )
-                                        .await;
-                                        if *discord_cover_resolving_for.peek() == song_key_for_spawn
-                                        {
-                                            discord_cover_url.set(resolved);
-                                        }
-                                    }
-                                    .instrument(tracing::info_span!("presence.cover_resolve")),
-                                );
-                            }
-                        }
-
-                        if discord_enabled {
-                            let song_changed = title != *last_title.peek();
-                            let resumed = !*was_playing.peek();
-                            let toggled_on = !last_discord_enabled;
-                            let cover_just_resolved =
-                                discord_cover_url.peek().is_some() && !*discord_cover_sent.peek();
-
-                            if song_changed
-                                || resumed
-                                || toggled_on
-                                || cover_just_resolved
-                                || source_changed
-                            {
-                                last_title.set(title.clone());
-                                last_source.set(discord_source_name.map(str::to_owned));
-
-                                let resolved = discord_cover_url.read().clone();
-                                let cover_ref = resolved.as_deref();
-
-                                let _ = p.set_now_playing(
-                                    &title,
-                                    &artist,
-                                    &album,
-                                    progress,
-                                    duration,
-                                    cover_ref,
-                                    discord_source_name,
-                                );
-
-                                if resolved.is_some() {
-                                    discord_cover_sent.set(true);
-                                }
-                            }
-                        } else if last_discord_enabled {
-                            let _ = p.clear_activity();
-                        }
-                    }
-
-                    // A transition is in flight from when a crossfade arms
-                    // (is_loading, while the next stream resolves) until its
-                    // deferred UI commits (pending_crossfade_ui). Arming or
-                    // skipping during that window double-fires: `pos` is the
-                    // incoming track's position while `duration` is still the
-                    // outgoing track's, so `remaining` is nonsense, and the
-                    // engine may still be finishing the outgoing track near its
-                    // end — which re-satisfies the arm and stacks transitions.
-                    let transition_in_flight =
-                        *ctrl.is_loading.read() || ctrl.pending_crossfade_ui.read().is_some();
-
-                    let remaining_secs = duration.saturating_sub(pos_secs);
-                    let should_crossfade = !external
-                        && duration > 0
-                        && pos_secs < duration
-                        && ctrl.should_crossfade()
-                        && ctrl.has_next_track()
-                        && remaining_secs <= config.read().crossfade_seconds as u64
-                        && *ctrl.armed_transition.peek() != Some(current_token);
-
-                    if should_crossfade && !transition_in_flight {
-                        ctrl.armed_transition.set(Some(current_token));
-                        {
-                            let mut config_write = config.write();
-                            let idx = *ctrl.current_queue_index.peek();
-                            if let Some(track) = ctrl.get_track_at(idx) {
-                                let track_id = track.id.uid().to_string();
-                                let source = ctrl.active_source.peek().clone();
-                                let count_key = source.source().listen_count_key(&track_id);
-                                *config_write.listen_counts.entry(count_key).or_insert(0) += 1;
-                                drop(config_write);
-                                bump_listen_count_db(track_id, source);
-                            }
-                        }
-                        ctrl.play_next_with_crossfade();
-                        nudge_event_loop();
-                        prev_playing = is_playing;
-                        {
-                            was_playing.set(is_playing);
-                            last_discord_enabled = discord_enabled;
-                        }
-                        continue;
-                    }
-
-                    let is_radio = duration == u64::MAX;
-                    let should_skip = if is_radio || external {
-                        false
-                    } else {
-                        ctrl.player.read().is_playback_complete()
-                            || (duration > 0 && pos.as_secs() >= duration.saturating_add(5))
-                    };
-
-                    if should_skip && !transition_in_flight {
-                        if !is_radio && duration > 0 && last_progress_secs != duration {
-                            last_progress_secs = duration;
-                            ctrl.current_song_progress.set(duration);
-                        }
-                        {
-                            let mut config_write = config.write();
-                            let _q = ctrl.queue.peek();
-                            let idx = *ctrl.current_queue_index.peek();
-                            if let Some(track) = ctrl.get_track_at(idx) {
-                                let track_id = track.id.uid().to_string();
-                                let source = ctrl.active_source.peek().clone();
-                                let count_key = source.source().listen_count_key(&track_id);
-                                *config_write.listen_counts.entry(count_key).or_insert(0) += 1;
-                                drop(config_write);
-                                bump_listen_count_db(track_id, source);
-                            }
-                        }
-                        ctrl.play_next();
-                        nudge_event_loop();
-                    }
-                } else {
-                    if *was_playing.peek() {
-                        if let Some(ref p) = presence {
-                            let title = ctrl.current_song_title.read().clone();
-                            let artist = ctrl.current_song_artist.read().clone();
-                            let album = ctrl.current_song_album.read().clone();
-                            if discord_enabled && discord_paused_enabled {
-                                let resolved = discord_cover_url.read().clone();
-                                last_source.set(discord_source_name.map(str::to_owned));
-                                let _ = p.set_paused(
-                                    &title,
-                                    &artist,
-                                    &album,
-                                    resolved.as_deref(),
-                                    discord_source_name,
-                                );
-                            } else if last_discord_enabled || !discord_paused_enabled {
-                                let _ = p.clear_activity();
-                            }
-                        }
-                    } else if let Some(ref p) = presence {
-                        if !discord_enabled && last_discord_enabled {
-                            let _ = p.clear_activity();
-                        } else if discord_enabled
-                            && (!last_discord_enabled || (source_changed && discord_paused_enabled))
-                        {
-                            let title = ctrl.current_song_title.read().clone();
-                            if !title.is_empty() {
-                                let artist = ctrl.current_song_artist.read().clone();
-                                let album = ctrl.current_song_album.read().clone();
-                                let resolved = discord_cover_url.read().clone();
-                                last_source.set(discord_source_name.map(str::to_owned));
-                                let _ = p.set_paused(
-                                    &title,
-                                    &artist,
-                                    &album,
-                                    resolved.as_deref(),
-                                    discord_source_name,
-                                );
-                            }
-                        }
-                    }
-                }
-
-                prev_playing = is_playing;
-                {
-                    was_playing.set(is_playing);
-                    last_discord_enabled = discord_enabled;
-                }
-            }
-        }
-    });
+    event_pump::use_player_event_pump(ctrl, config);
 }
